+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Unreleased
 
-### Added
-
+-   [Patch] Simplify `<noscript>` output in `Image` component by removing `<picture>` and `srcSet` and using only `<img alt="..." src="..." />`.
 -   [Minor] Add the new `Popover` component.
 
 ## 6.2.1 - 2019-06-24